requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements-dev.in
#
asgiref==3.6.0
    # via django
django==4.2
    # via
    #   -r requirements-dev.in
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==3.14.0
    # via -r requirements-dev.in
django-filter==23.2
    # via -r requirements-dev.in
djangorestframework==3.14.0
    # via
    #   -r requirements-dev.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.2
    # via -r requirements-dev.in
pyjwt==2.6.0
    # via djangorestframework-simplejwt
pytz==2023.3
    # via djangorestframework
sqlparse==0.4.4
    # via django
tzdata==2023.3
    # via django
